Fix continuous creation of empty ghost entries when attempting to create a new entry
The user object seems to receive continuous updates and thus cannot be listed in the dependency array directly. Instead, list the `user.partnerIds` member that is actually used in the event hook. This behavior was reported in https://github.com/ImmunHelden/ImmunHelden.de/issues/201 and should be fixed once this change gets deployed.